Remove unnecessary file `util.ts`. (#1170)
Remove duplicate file util.ts that contained VersionsData
